20190129:  - Cori doesn't like a single session to open and close the same file >10,000 times. Now, after creating all the hdf5 groups and creating and sizing all the hdf5 datasets, so that the file metadata will not change, I create and leave open an h5py file handler in serial read-only mode on all MPI ranks. This appears to work even when intermittently opening a second file handler to the same file in collective mpio write mode and modifying values in existing datasets.

+++ optimize_DG_GC_synaptic_integration.py
@@ -327,7 +327,7 @@ def shutdown_worker():
 
     """
     context.open_temp_model_data_file.close()
-    if context.interface.global_comm.rank == 0:
+    if context.interface.global_comm.rank == 0 and not context.debug:
         os.remove(context.temp_model_data_file_path)
 
 
@@ -979,6 +979,12 @@ def get_objectives_synaptic_integration(features, export=False):
                 for syn_condition in soma_compound_EPSP_amp[syn_group]:
                     this_group.create_dataset(syn_condition, compression='gzip',
                                               data=soma_compound_EPSP_amp[syn_group][syn_condition])
+
+    if context.debug:
+        print 'features:'
+        pprint.pprint(features)
+        print 'objectives:'
+        pprint.pprint(objectives)
 
     return features, objectives
